db-adapter/requirements.txt
mysql-connector-python
paho-mqtt
influxdb